Ionis kutrolli development: User authentication for commenting and comment deletion.

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteCommentServlet.java
@@ -4,6 +4,11 @@ import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
+import com.google.appengine.api.datastore.Entity;
+import com.google.appengine.api.datastore.EntityNotFoundException;
+
 import java.io.IOException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -11,17 +16,30 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
- * Servlet for handleing the deletion of one selected comment
+ * Servlet for handling the deletion of one selected comment
  */
 @WebServlet("/delete-comment")
 public class DeleteCommentServlet extends HttpServlet {
 
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    long id = Long.parseLong(request.getParameter("id"));
+    UserService userService = UserServiceFactory.getUserService();
+    if (userService.isUserLoggedIn()) {
+    
+      long id = Long.parseLong(request.getParameter("id"));
 
-    Key commentEntityKey = KeyFactory.createKey("Comment", id);
-    DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-    datastore.delete(commentEntityKey);
+      Key commentEntityKey = KeyFactory.createKey("Comment", id);
+      DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
+      Entity commentEntity;
+      try {
+        commentEntity = datastore.get(commentEntityKey);
+      } catch (EntityNotFoundException e) {
+        System.out.println("Cannot find comment to delete.");
+        return;
+      }
+      if (userService.getCurrentUser().getUserId().equals(commentEntity.getProperty("userId"))) {
+        datastore.delete(commentEntityKey);
+      }
+    }
   }
 }
